requirements.txt
fastapi
uvicorn
Pillow
pytesseract
opencv-python-headless
numpy